+++ src/components/Modals/index.ts
@@ -2,6 +2,7 @@
 import AddTaskModal from './AddTaskModal';
 import BreakModal from './BreakModal';
 import InfoModal from './InfoModal';
+import AddSubtaskModal from './AddSubtaskModal';
 
-export { BreakModal, InfoModal };
+export { BreakModal, InfoModal, AddSubtaskModal };
 export default AddTaskModal;